fix cannot use _Ctype_longlong(size) (value of type _Ctype_longlong) as _Ctype_Py_ssize_t value in argument to (_Cfunc_PyBytes_FromStringAndSize) with int64(size) fix

+++ etlx.go
@@ -449,7 +449,7 @@ func Slice_byte_to_bytes(handle CGoHandle) *C.PyObject {
 	s := deptrFromHandle_Slice_byte(handle)
 	ptr := unsafe.Pointer(&s[0])
 	size := len(s)
-	return C.PyBytes_FromStringAndSize((*C.char)(ptr), C.longlong(size))
+	return C.PyBytes_FromStringAndSize((*C.char)(ptr), C.longlong(int64(size)))
 }
 
 // Converters for implicit pointer handles for type: []error
